# public/swagger.json
{"openapi":"3.0.0","paths":{"/stripe/customers/create":{"post":{"operationId":"StripeCustomerController_createStripeCustomer","parameters":[],"requestBody":{"required":true,"description":"Stripe Customer details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["stripe-customers"]}},"/stripe/customers/update":{"put":{"operationId":"StripeCustomerController_updateStripeCustomer","parameters":[],"requestBody":{"required":true,"description":"Update Stripe Customer details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerDto"}}}},"responses":{"200":{"description":"Stripe customer updated successfully"}},"security":[{"bearer":[]}],"tags":["stripe-customers"]}},"/stripe/customers":{"get":{"operationId":"StripeCustomerController_getStripeCustomer","parameters":[],"responses":{"200":{"description":"Stripe customer retrieved successfully"}},"security":[{"bearer":[]}],"tags":["stripe-customers"]},"delete":{"operationId":"StripeCustomerController_deleteCustomer","parameters":[],"responses":{"200":{"description":"Stripe customer deleted successfully"},"400":{"description":"Error deleting stripe customer"},"404":{"description":"Stripe customer not found"}},"security":[{"bearer":[]}],"tags":["stripe-customers"]}},"/stripe/customers/payment-methods":{"get":{"operationId":"StripeCustomerController_listPaymentMethods","parameters":[],"responses":{"200":{"description":"Payment methods retrieved successfully."}},"security":[{"bearer":[]}],"summary":"List authenticated user's saved payment methods (cards)","tags":["stripe-customers"]}},"/stripe/customers/setup-intent":{"post":{"operationId":"StripeCustomerController_createSetupIntent","parameters":[],"responses":{"201":{"description":"SetupIntent created successfully. Returns client_secret."}},"security":[{"bearer":[]}],"summary":"Create a SetupIntent to add a new payment method","tags":["stripe-customers"]}},"/stripe/customers/payment-methods/set-default":{"put":{"operationId":"StripeCustomerController_setDefaultPaymentMethod","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetDefaultPaymentMethodDto"}}}},"responses":{"200":{"description":"Default payment method updated successfully."}},"security":[{"bearer":[]}],"summary":"Set a default payment method for the authenticated user","tags":["stripe-customers"]}},"/stripe/customers/payment-methods/{pmId}":{"delete":{"operationId":"StripeCustomerController_detachPaymentMethod","parameters":[{"name":"pmId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Payment method detached successfully."}},"security":[{"bearer":[]}],"summary":"Detach (delete) a saved payment method","tags":["stripe-customers"]}},"/stripe/subscriptions/subscribe":{"post":{"operationId":"SubscriptionController_createSubscription","parameters":[],"requestBody":{"required":true,"description":"Details of the subscription plan to create","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubscriptionDto"}}}},"responses":{"201":{"description":"Subscription created or initiated successfully."},"400":{"description":"Invalid input, payment method issue, or other creation error."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Create a new Stripe subscription for the authenticated user","tags":["stripe-subscriptions"]}},"/stripe/subscriptions/change-plan":{"put":{"operationId":"SubscriptionController_changePlan","parameters":[],"requestBody":{"required":true,"description":"Details of the new plan to switch to.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePlanDto"}}}},"responses":{"200":{"description":"Subscription plan change initiated successfully. Check status for outcome."},"400":{"description":"Invalid input, no active subscription, or other update error."},"404":{"description":"User or active subscription not found."}},"security":[{"bearer":[]}],"summary":"Change (upgrade/downgrade) the authenticated user's active subscription plan","tags":["stripe-subscriptions"]}},"/stripe/subscriptions/current":{"get":{"operationId":"SubscriptionController_getCurrentSubscription","parameters":[{"name":"sync","required":false,"in":"query","description":"Whether to sync with Stripe for the latest data (default: false)","schema":{"type":"boolean"}}],"responses":{"200":{"description":"Subscription details retrieved successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionResponseDto"}}}},"400":{"description":"Error retrieving subscription details."},"404":{"description":"User not found."}},"security":[{"bearer":[]}],"summary":"Get the authenticated user's current subscription details","tags":["stripe-subscriptions"]}},"/stripe/subscriptions/cancel":{"delete":{"operationId":"SubscriptionController_cancelSubscription","parameters":[],"requestBody":{"required":false,"description":"Details of the cancellation request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelSubscriptionDto"}}}},"responses":{"200":{"description":"Subscription cancellation initiated successfully. Check status for outcome."},"400":{"description":"Invalid input, no active subscription, or other cancellation error."},"404":{"description":"User or active subscription not found."}},"security":[{"bearer":[]}],"summary":"Cancel the authenticated user's active subscription (if any) immediately","tags":["stripe-subscriptions"]}},"/api/internal/subscription/subscription-details/{userId}":{"get":{"operationId":"InternalSubscriptionController_getUserSubscriptionPlan","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["InternalSubscription"]}},"/wallet/top-up":{"post":{"operationId":"WalletController_topUp","parameters":[{"name":"idempotency-key","required":true,"in":"header","description":"Unique key to prevent duplicate top-up requests","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletTopUpDto"}}}},"responses":{"200":{"description":"Wallet top-up successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletTopUpResponseDto"}}}},"400":{"description":"Bad Request (e.g., card declined)"},"500":{"description":"Internal Server Error"}},"summary":"Top up the user wallet balance","tags":["Wallet"]}},"/wallet/balance":{"get":{"description":"Retrieve the current wallet balance for the authenticated user","operationId":"WalletController_getWalletBalance","parameters":[],"responses":{"200":{"description":"Successfully retrieved wallet balance","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"_id":{"type":"string","example":"507f1f77bcf86cd799439011"},"userId":{"type":"string","example":"507f1f77bcf86cd799439012"},"balance":{"type":"number","example":150000,"description":"Balance in cents"},"currency":{"type":"string","example":"USD"},"status":{"type":"string","enum":["ACTIVE","FROZEN","CLOSED"],"example":"ACTIVE"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"message":{"type":"string","example":"Successfully fetched user's balance"},"success":{"type":"boolean","example":true}}}}}},"401":{"description":"Unauthorized - Invalid or missing authentication"},"404":{"description":"User not found"},"500":{"description":"Internal Server Error"}},"summary":"Get user wallet balance","tags":["Wallet"]}},"/api/internal/wallet/debit-for-campaign":{"post":{"operationId":"InternalWalletController_debitWalletForCampaign","parameters":[{"name":"idempotency-key","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DebitCampaignDto"}}}},"responses":{"201":{"description":""}},"tags":["InternalWallet"]}}},"info":{"title":"Amplify Wallet","description":"Amplify Wallet API for handling payments","version":"1.0","contact":{}},"tags":[{"name":"customers","description":"Stripe Customers endpoint"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateCustomerDto":{"type":"object","properties":{"metadata":{"type":"object","description":"The address of the customer","example":{"line1":"123 Main St","city":"Anytown","state":"CA","postal_code":"12345","country":"US"}},"cardHolderName":{"type":"string","description":"The name of the card holder","example":"John Doe"}}},"AddressDto":{"type":"object","properties":{"line1":{"type":"string"},"line2":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"postal_code":{"type":"string"},"country":{"type":"string"}}},"ShippingDto":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"}},"required":["address"]},"UpdateCustomerDto":{"type":"object","properties":{"phone":{"type":"string","description":"Customer phone number"},"metadata":{"type":"object","description":"Customer metadata"},"defaultPaymentMethod":{"type":"string","description":"Default payment method ID"},"address":{"$ref":"#/components/schemas/AddressDto"},"shipping":{"$ref":"#/components/schemas/ShippingDto"},"description":{"type":"string","description":"Customer description"}}},"SetDefaultPaymentMethodDto":{"type":"object","properties":{"paymentMethodId":{"type":"string","description":"The Stripe PaymentMethod ID (pm_...) to set as default."}},"required":["paymentMethodId"]},"CreateSubscriptionDto":{"type":"object","properties":{"priceId":{"type":"string","description":"The Stripe Price ID for the desired base subscription plan (e.g., price_starter_monthly)","example":"price_starter_monthly"},"paymentMethodId":{"type":"string","description":"The Stripe Payment Method ID (required for the first paid subscription, obtained from frontend Stripe Elements/SDK)","example":"pm_1P7kvL2eZvKYlo2C9GvA3B4C"}},"required":["priceId"]},"ChangePlanDto":{"type":"object","properties":{"newPriceId":{"type":"string","description":"The Stripe Price ID for the new desired subscription plan (e.g., price_grow_monthly).","example":"price_grow_monthly_actual_id"},"prorationBehavior":{"type":"string","description":"Determines how to handle prorations when switching plans. Defaults to 'create_prorations'.","example":"create_prorations","enum":["create_prorations","none","always_invoice"]}},"required":["newPriceId"]},"SubscriptionResponseDto":{"type":"object","properties":{"stripeSubscriptionId":{"type":"string","description":"Stripe subscription ID"},"subscriptionStatus":{"type":"string","description":"Current subscription status"},"activeStripePriceId":{"type":"string","description":"Active Stripe price ID"},"currentPeriodEnd":{"format":"date-time","type":"string","description":"Current billing period end date"},"hasActiveSubscription":{"type":"boolean","description":"Whether user has an active subscription"},"paymentStatus":{"type":"string","description":"Current payment status"},"defaultPaymentMethod":{"type":"string","description":"Default payment method ID"},"lastStripeSync":{"format":"date-time","type":"string","description":"Last sync with Stripe"}},"required":["stripeSubscriptionId","subscriptionStatus","activeStripePriceId","currentPeriodEnd","hasActiveSubscription","paymentStatus","defaultPaymentMethod","lastStripeSync"]},"CancelSubscriptionDto":{"type":"object","properties":{"cancelImmediately":{"type":"boolean","description":"Whether to cancel immediately or at the end of the current period","default":false},"cancellationReason":{"type":"string","description":"Optional reason for cancellation (for internal tracking)"}}},"WalletTopUpDto":{"type":"object","properties":{"amount":{"type":"number","description":"The amount to top up the wallet","example":100},"paymentMethodId":{"type":"string","description":"The Stripe Payment Method ID (required for the first paid subscription, obtained from frontend Stripe Elements/SDK)","example":"pm_1P7kvL2eZvKYlo2C9GvA3B4C"}},"required":["amount"]},"Transaction":{"type":"object","properties":{"userId":{"type":"string","description":"The ID of the user associated with the transaction","example":"60c72b2f9b1e8b0015f8e5b1"},"type":{"type":"string","enum":["TOP_UP","CAMPAIGN_DEBIT","REFUND"],"description":"The type of transaction (e.g., TOP_UP, CAMPAIGN_DEBIT)","example":"TOP_UP"},"amount":{"type":"number","description":"The amount of the transaction in cents","example":10000},"status":{"type":"string","enum":["PENDING","COMPLETED","FAILED"],"description":"The current status of the transaction","example":"COMPLETED"},"idempotencyKey":{"type":"string","description":"Unique key to prevent duplicate transactions (for client-initiated actions)","example":"your_unique_idempotency_key_123"},"metadata":{"type":"object","description":"Additional metadata related to the transaction","additionalProperties":true,"example":{"paymentIntentId":"pi_abc123","someOtherDetail":"value"}}},"required":["userId","type","amount","status","metadata"]},"WalletTopUpResponseDto":{"type":"object","properties":{"data":{"description":"The completed transaction document.","additionalProperties":true,"allOf":[{"$ref":"#/components/schemas/Transaction"}]},"message":{"type":"string","description":"Response message","example":"Operation completed successfully"},"success":{"type":"boolean","description":"Response status","example":"success"}},"required":["data","message","success"]},"DebitCampaignDto":{"type":"object","properties":{"userId":{"type":"string","description":"The ID of the user to debit","example":"9182000182fedc"}},"required":["userId"]}}}}